Move classes into packages
Packaging is good for organisation.

Another thing we need to do while packaging is to
update runtest.bat to compile code within the packages.
