+++ src/main/java/duke/CommandMatcher.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.util.function.Consumer;
 import java.util.function.Predicate;
