+++ src/main/java/duke/PrefixCommandMatcher.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.util.HashMap;
 import java.util.Map;
 import java.util.function.BiConsumer;
