+++ src/main/java/duke/task/Deadline.java
@@ -1,3 +1,7 @@
+package duke.task;
+
+import duke.ParsedDateTime;
+
 /**
  * Handles a task with a deadline.
  */
@@ -5,7 +9,7 @@ public class Deadline extends Task {
     protected ParsedDateTime datetime;
 
     /**
-     * Creates a Deadline object.
+     * Creates a duke.tasks.Deadline object.
      * @param description Description of deadline.
      * @param by Time of deadline.
      */
@@ -14,7 +18,7 @@ public class Deadline extends Task {
     }
 
     /**
-     * Creates a Deadline object.
+     * Creates a duke.tasks.Deadline object.
      * @param description Description of deadline.
      * @param by Time of deadline.
      * @param done If the task is done.
@@ -36,6 +40,6 @@ public class Deadline extends Task {
     @Override
     public String[] getAsStringArray() {
         String[] data = super.getAsStringArray();
-        return new String[]{ "Deadline", data[1], data[2], datetime.toString() };
+        return new String[]{ "duke.tasks.Deadline", data[1], data[2], datetime.toString() };
     }
 }
