+++ src/main/java/duke/task/TaskStorage.java
@@ -1,6 +1,10 @@
+package duke.task;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
+
+import duke.Storage;
 
 /**
  * Accesses a file for tasks.
