+++ text-ui-test/duke.txt
@@ -1,16 +1,16 @@
-/v8AVABvAEQAbw==,/v8AaABvAG0AZQB3AG8AcgBr,/v8AZgBhAGwAcwBl
-/v8ARQB2AGUAbgB0,/v8AbQBlAGUAdAAgAHAAZQBvAHAAbABlACAAYQB0ACAAZgB1AG4AYwB0AGkAbwBuACAAeAAvAHkALwB6ACAAdwBoAGEAdABlAHYAZQBy,/v8AdAByAHUAZQ==,/v8ANwA6ADEAMgBhAG0=
-/v8ARABlAGEAZABsAGkAbgBl,/v8ALwBhAHQAIABtAHkAcwB0AGUAcgBpAG8AdQBz,/v8AdAByAHUAZQ==,/v8AWwB1AG4AawBuAG8AdwBuAF0=
-/v8ARABlAGEAZABsAGkAbgBl,/v8ALwBiAHkAIABtAHkAcwB0AGUAcgBpAG8AdQBzACAAbgB1AG0AYgBlAHIAIAAy,/v8AdAByAHUAZQ==,/v8AWwB1AG4AawBuAG8AdwBuAF0=
-/v8ARABlAGEAZABsAGkAbgBl,/v8AcgB1AHMAaAAgAHQAaABpAHMAIABwAGwAZQBhAHMAZQ==,/v8AZgBhAGwAcwBl,/v8AdwBoAGUAbgA/
-/v8ARQB2AGUAbgB0,/v8AdwBvAHcAIAAgAG0AdQBjAGgAIAAgAHMAcABhAGMAZQ==,/v8AZgBhAGwAcwBl,/v8AMQAwAGEAbQ==
-/v8AVABvAEQAbw==,/v8AaABvAG0AZQB3AG8AcgBr,/v8AZgBhAGwAcwBl
-/v8AVABvAEQAbw==,/v8AbwB0AGgAZQByACAAaABvAG0AZQB3AG8AcgBr,/v8AZgBhAGwAcwBl
-/v8AVABvAEQAbw==,/v8AcAByAG8AagBlAGMAdAAgADEAMg==,/v8AZgBhAGwAcwBl
-/v8ARQB2AGUAbgB0,/v8AbQBlAGUAdAAgAGYAcgBpAGUAbgBkAHM=,/v8AZgBhAGwAcwBl,/v8AWwB1AG4AawBuAG8AdwBuAF0=
-/v8ARQB2AGUAbgB0,/v8AbQBlAGUAdAAgAHAAZQBvAHAAbABlACAAYQB0ACAAZABpAG4AbgBlAHI=,/v8AZgBhAGwAcwBl,/v8ANgA6ADIANQAgAHAAbQ==
-/v8ARQB2AGUAbgB0,/v8AbQBlAGUAdAAgAHAAZQBvAHAAbABlACAAYQB0ACAAZgB1AG4AYwB0AGkAbwBuACAAeAAvAHkALwB6ACAAdwBoAGEAdABlAHYAZQBy,/v8AZgBhAGwAcwBl,/v8ANwA6ADEAMgBhAG0=
-/v8ARABlAGEAZABsAGkAbgBl,/v8ALwBhAHQAIABtAHkAcwB0AGUAcgBpAG8AdQBz,/v8AZgBhAGwAcwBl,/v8AWwB1AG4AawBuAG8AdwBuAF0=
-/v8ARABlAGEAZABsAGkAbgBl,/v8ALwBiAHkAIABtAHkAcwB0AGUAcgBpAG8AdQBzACAAbgB1AG0AYgBlAHIAIAAy,/v8AZgBhAGwAcwBl,/v8AWwB1AG4AawBuAG8AdwBuAF0=
-/v8ARABlAGEAZABsAGkAbgBl,/v8AcgB1AHMAaAAgAHQAaABpAHMAIABwAGwAZQBhAHMAZQ==,/v8AZgBhAGwAcwBl,/v8AdwBoAGUAbgA/
-/v8ARQB2AGUAbgB0,/v8AdwBvAHcAIAAgAG0AdQBjAGgAIAAgAHMAcABhAGMAZQ==,/v8AZgBhAGwAcwBl,/v8AMQAwAGEAbQ==
+/v8AZAB1AGsAZQAuAHQAYQBzAGsAcwAuAFQAbwBEAG8=,/v8AaABvAG0AZQB3AG8AcgBr,/v8AZgBhAGwAcwBl
+/v8AZAB1AGsAZQAuAHQAYQBzAGsAcwAuAEUAdgBlAG4AdA==,/v8AbQBlAGUAdAAgAHAAZQBvAHAAbABlACAAYQB0ACAAZgB1AG4AYwB0AGkAbwBuACAAeAAvAHkALwB6ACAAdwBoAGEAdABlAHYAZQBy,/v8AdAByAHUAZQ==,/v8ANwA6ADEAMgBhAG0=
+/v8AZAB1AGsAZQAuAHQAYQBzAGsAcwAuAEQAZQBhAGQAbABpAG4AZQ==,/v8ALwBhAHQAIABtAHkAcwB0AGUAcgBpAG8AdQBz,/v8AdAByAHUAZQ==,/v8AWwB1AG4AawBuAG8AdwBuAF0=
+/v8AZAB1AGsAZQAuAHQAYQBzAGsAcwAuAEQAZQBhAGQAbABpAG4AZQ==,/v8ALwBiAHkAIABtAHkAcwB0AGUAcgBpAG8AdQBzACAAbgB1AG0AYgBlAHIAIAAy,/v8AdAByAHUAZQ==,/v8AWwB1AG4AawBuAG8AdwBuAF0=
+/v8AZAB1AGsAZQAuAHQAYQBzAGsAcwAuAEQAZQBhAGQAbABpAG4AZQ==,/v8AcgB1AHMAaAAgAHQAaABpAHMAIABwAGwAZQBhAHMAZQ==,/v8AZgBhAGwAcwBl,/v8AdwBoAGUAbgA/
+/v8AZAB1AGsAZQAuAHQAYQBzAGsAcwAuAEUAdgBlAG4AdA==,/v8AdwBvAHcAIAAgAG0AdQBjAGgAIAAgAHMAcABhAGMAZQ==,/v8AZgBhAGwAcwBl,/v8AMQAwAGEAbQ==
+/v8AZAB1AGsAZQAuAHQAYQBzAGsAcwAuAFQAbwBEAG8=,/v8AaABvAG0AZQB3AG8AcgBr,/v8AZgBhAGwAcwBl
+/v8AZAB1AGsAZQAuAHQAYQBzAGsAcwAuAFQAbwBEAG8=,/v8AbwB0AGgAZQByACAAaABvAG0AZQB3AG8AcgBr,/v8AZgBhAGwAcwBl
+/v8AZAB1AGsAZQAuAHQAYQBzAGsAcwAuAFQAbwBEAG8=,/v8AcAByAG8AagBlAGMAdAAgADEAMg==,/v8AZgBhAGwAcwBl
+/v8AZAB1AGsAZQAuAHQAYQBzAGsAcwAuAEUAdgBlAG4AdA==,/v8AbQBlAGUAdAAgAGYAcgBpAGUAbgBkAHM=,/v8AZgBhAGwAcwBl,/v8AWwB1AG4AawBuAG8AdwBuAF0=
+/v8AZAB1AGsAZQAuAHQAYQBzAGsAcwAuAEUAdgBlAG4AdA==,/v8AbQBlAGUAdAAgAHAAZQBvAHAAbABlACAAYQB0ACAAZABpAG4AbgBlAHI=,/v8AZgBhAGwAcwBl,/v8ANgA6ADIANQAgAHAAbQ==
+/v8AZAB1AGsAZQAuAHQAYQBzAGsAcwAuAEUAdgBlAG4AdA==,/v8AbQBlAGUAdAAgAHAAZQBvAHAAbABlACAAYQB0ACAAZgB1AG4AYwB0AGkAbwBuACAAeAAvAHkALwB6ACAAdwBoAGEAdABlAHYAZQBy,/v8AZgBhAGwAcwBl,/v8ANwA6ADEAMgBhAG0=
+/v8AZAB1AGsAZQAuAHQAYQBzAGsAcwAuAEQAZQBhAGQAbABpAG4AZQ==,/v8ALwBhAHQAIABtAHkAcwB0AGUAcgBpAG8AdQBz,/v8AZgBhAGwAcwBl,/v8AWwB1AG4AawBuAG8AdwBuAF0=
+/v8AZAB1AGsAZQAuAHQAYQBzAGsAcwAuAEQAZQBhAGQAbABpAG4AZQ==,/v8ALwBiAHkAIABtAHkAcwB0AGUAcgBpAG8AdQBzACAAbgB1AG0AYgBlAHIAIAAy,/v8AZgBhAGwAcwBl,/v8AWwB1AG4AawBuAG8AdwBuAF0=
+/v8AZAB1AGsAZQAuAHQAYQBzAGsAcwAuAEQAZQBhAGQAbABpAG4AZQ==,/v8AcgB1AHMAaAAgAHQAaABpAHMAIABwAGwAZQBhAHMAZQ==,/v8AZgBhAGwAcwBl,/v8AdwBoAGUAbgA/
+/v8AZAB1AGsAZQAuAHQAYQBzAGsAcwAuAEUAdgBlAG4AdA==,/v8AdwBvAHcAIAAgAG0AdQBjAGgAIAAgAHMAcABhAGMAZQ==,/v8AZgBhAGwAcwBl,/v8AMQAwAGEAbQ==
