+++ text-ui-test/runtest.bat
@@ -7,7 +7,7 @@ REM delete output from previous run
 if exist ACTUAL.TXT del ACTUAL.TXT
 
 REM compile the code into the bin folder
-javac  -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\*.java
+javac  -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\duke\*.java ..\src\main\java\duke\task\*.java
 IF ERRORLEVEL 1 (
     echo ********** BUILD FAILURE **********
     exit /b 1
@@ -18,7 +18,7 @@ REM delete data
 if exist duke.txt del duke.txt
 
 REM run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ..\bin Duke < input.txt > ACTUAL.TXT
+java -classpath ..\bin duke.Duke < input.txt > ACTUAL.TXT
 
 REM compare the output to the expected output
 FC ACTUAL.TXT EXPECTED.TXT
@@ -27,7 +27,7 @@ REM delete output from previous run
 if exist ACTUAL1.TXT del ACTUAL1.TXT
 
 REM run the program, feed commands from input.txt file and redirect the output to the ACTUAL1.TXT
-java -classpath ..\bin Duke < input.txt > ACTUAL1.TXT
+java -classpath ..\bin duke.Duke < input.txt > ACTUAL1.TXT
 
 REM compare the output to the expected output
 FC ACTUAL1.TXT EXPECTED1.TXT
